+++ src/global_artlist.js
@@ -1,5 +1,6 @@
 const png = ".png"
 const jpeg = ".jpeg"
+const DEFAULT_DIMENSIONS = [500]
 
 export const ARTLIST = []
 
@@ -11,29 +12,29 @@ function addArtwork(img = "test/test1", dimensions = [400,400], ext = png) {
     })
 }
 
-addArtwork('ADHD!', [200]);
-addArtwork('AHH', [200]);
-addArtwork('And you say you have no fire powers right', [200]);
-addArtwork('barakah sticcks', [200])
-addArtwork('clones', [200])
-addArtwork('failure', [200])
-addArtwork('GTA VI', [200])
-addArtwork('hey!', [200])
-addArtwork("romantic", [200])
-addArtwork("math", [200])
-addArtwork('Jump!', [200])
-addArtwork('lies', [200])
-addArtwork('Magic is a big secret', [200])
-addArtwork('mining with the boys', [200], jpeg)
-addArtwork('POWER', [200])
-addArtwork('pyromaniac', [200])
-addArtwork('rate my fit', [200])
-addArtwork('shootergame', [200])
-addArtwork('SnowingSeason', [200])
-addArtwork('The Mountains', [200], jpeg)
-addArtwork('the obs ', [200])
-addArtwork('what movie are you here to see', [200])
-addArtwork('what6x3', [200])
-addArtwork('whatyagotthere', [200])
-addArtwork('Wrestle', [200])
-addArtwork('yournotsick ', [200])
+addArtwork('ADHD!', DEFAULT_DIMENSIONS);
+addArtwork('AHH', DEFAULT_DIMENSIONS);
+addArtwork('And you say you have no fire powers right', DEFAULT_DIMENSIONS);
+addArtwork('barakah sticcks', DEFAULT_DIMENSIONS)
+addArtwork('clones', DEFAULT_DIMENSIONS)
+addArtwork('failure', DEFAULT_DIMENSIONS)
+addArtwork('GTA VI', DEFAULT_DIMENSIONS)
+addArtwork('hey!', DEFAULT_DIMENSIONS)
+addArtwork("romantic", DEFAULT_DIMENSIONS)
+addArtwork("math", DEFAULT_DIMENSIONS)
+addArtwork('Jump!', DEFAULT_DIMENSIONS)
+addArtwork('lies', DEFAULT_DIMENSIONS)
+addArtwork('Magic is a big secret', DEFAULT_DIMENSIONS)
+addArtwork('mining with the boys', DEFAULT_DIMENSIONS, jpeg)
+addArtwork('POWER', DEFAULT_DIMENSIONS)
+addArtwork('pyromaniac', DEFAULT_DIMENSIONS)
+addArtwork('rate my fit', DEFAULT_DIMENSIONS)
+addArtwork('shootergame', DEFAULT_DIMENSIONS)
+addArtwork('SnowingSeason', DEFAULT_DIMENSIONS)
+addArtwork('The Mountains', DEFAULT_DIMENSIONS, jpeg)
+addArtwork('the obs ', DEFAULT_DIMENSIONS)
+addArtwork('what movie are you here to see', DEFAULT_DIMENSIONS)
+addArtwork('what6x3', DEFAULT_DIMENSIONS)
+addArtwork('whatyagotthere', DEFAULT_DIMENSIONS)
+addArtwork('Wrestle', DEFAULT_DIMENSIONS)
+addArtwork('yournotsick ', DEFAULT_DIMENSIONS)