+++ CHANGES.md
@@ -3,6 +3,8 @@
 - Added 26 art images
 - Added function in `global_artwork.js` for adding artwork
 - Added jpeg constant var in `global_artwork.js` file
+- Added default dimensions variable in the artwork file
+- Added Artwork filename text to artwork
 ## Removed
 - Removed 2 test imgs
 # v1h - 1 Hour Version - 10/23/2024